Add numeric literal-context inference + frontend diagnostic
Threads expected-type context through numeric literals across lexer,
typechecker (call/enum/struct/function checking) and C codegen; adds
literal_context test fixtures and frontend diagnostic coverage.

+++ src/ast/expressions.rs
@@ -12,7 +12,7 @@ pub use parts::{MatchArm, StringPart};
 #[derive(Debug, Clone, PartialEq, Serialize)]
 pub enum Expression {
     IntLiteral {
-        value: i64,
+        value: i128,
         span: Span,
     },
     FloatLiteral {
